Quite some changes
void resize_stack -> int resize_brainfuck_stack
Now returns 0 on success and 1 on error (instead of aborting)
On error, the old stack (if any) is unchanged.
The whole stack is now zero'd  when this function is called.

Added void free_brainfuck stack which frees and nulls the stack

void brainfuck() -> int brainfuck()
Returns 0 on success
Added a wrapper which checks for brackets before executing